7699 use mixin for model
