+++ netbox/virtualization/models/clusters.py
@@ -6,7 +6,7 @@ from django.utils.translation import gettext_lazy as _
 
 from dcim.models import Device
 from netbox.models import OrganizationalModel, PrimaryModel
-from netbox.models.features import ContactsMixin
+from netbox.models.features import CachedLocationScopeMixin, ContactsMixin
 from virtualization.choices import *
 from virtualization.constants import CLUSTER_SCOPE_TYPES
 
@@ -44,7 +44,7 @@ class ClusterGroup(ContactsMixin, OrganizationalModel):
         verbose_name_plural = _('cluster groups')
 
 
-class Cluster(ContactsMixin, PrimaryModel):
+class Cluster(ContactsMixin, CachedLocationScopeMixin, PrimaryModel):
     """
     A cluster of VirtualMachines. Each Cluster may optionally be associated with one or more Devices.
     """
@@ -103,36 +103,6 @@ class Cluster(ContactsMixin, PrimaryModel):
         related_query_name='cluster'
     )
 
-    # Cached associations to enable efficient filtering
-    _location = models.ForeignKey(
-        to='dcim.Location',
-        on_delete=models.CASCADE,
-        related_name='_clusters',
-        blank=True,
-        null=True
-    )
-    _site = models.ForeignKey(
-        to='dcim.Site',
-        on_delete=models.CASCADE,
-        related_name='_clusters',
-        blank=True,
-        null=True
-    )
-    _region = models.ForeignKey(
-        to='dcim.Region',
-        on_delete=models.CASCADE,
-        related_name='_clusters',
-        blank=True,
-        null=True
-    )
-    _sitegroup = models.ForeignKey(
-        to='dcim.SiteGroup',
-        on_delete=models.CASCADE,
-        related_name='_clusters',
-        blank=True,
-        null=True
-    )
-
     clone_fields = (
         'scope_type', 'scope_id', 'type', 'group', 'status', 'tenant',
     )
@@ -180,28 +150,3 @@ class Cluster(ContactsMixin, PrimaryModel):
                         "{count} devices are assigned as hosts for this cluster but are not in site {site}"
                     ).format(count=nonsite_devices, site=site)
                 })
-
-    def save(self, *args, **kwargs):
-        # Cache objects associated with the terminating object (for filtering)
-        self.cache_related_objects()
-
-        super().save(*args, **kwargs)
-
-    def cache_related_objects(self):
-        self._region = self._sitegroup = self._site = self._location = None
-        if self.scope_type:
-            scope_type = self.scope_type.model_class()
-            if scope_type == apps.get_model('dcim', 'region'):
-                self._region = self.scope
-            elif scope_type == apps.get_model('dcim', 'sitegroup'):
-                self._sitegroup = self.scope
-            elif scope_type == apps.get_model('dcim', 'site'):
-                self._region = self.scope.region
-                self._sitegroup = self.scope.group
-                self._site = self.scope
-            elif scope_type == apps.get_model('dcim', 'location'):
-                self._region = self.scope.site.region
-                self._sitegroup = self.scope.site.group
-                self._site = self.scope.site
-                self._location = self.scope
-    cache_related_objects.alters_data = True
